Rename the library from QRCode to QRCodeGenerator
