+++ src/QRCodeGenerator.h
